--- static/help/length.html
@@ -1,22 +1,0 @@
-<!-- Modal -->
-<div class="modal fade" id="length_modal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="exampleModalLabel">Pendulum length</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <div class="modal-body">
-                <p>The procedure to measure the string is described in String measurement. Then,
-                    the radius of the sphere/ball (in the vertical direction, i.e., along the string direction) should
-                    be added to the string length, resulting in the pendulum length. The unit of length for this
-                    quantity is meter (unlike the other ones which is centimeter). Once the pendulum length has been
-                    measured, store the value in the micro-controller flash by using the command "set pendulum length".
-                </p>
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-            </div>
-        </div>
-    </div>
-</div>
