--- static/help/max_pos.html
@@ -1,22 +1,0 @@
-<!-- Modal -->
-<div class="modal fade" id="max_pos_modal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-    <div class="modal-dialog modal-lg">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="exampleModalLabel">Maximum position</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <div class="modal-body">
-                <p>The shovel maximum position is measured by slowly displacing the shovel with
-                    the hand towards the end of the sliding structure (when the slider is fully extended). Once the
-                    shovel reaches the end position (cannot move further) record its absolute position by reading the
-                    value at the scale fixed on the launcher with the help of the indicator attached to the shovel.</p>
-                <br><img alt="Maximum position" src="/static/images/help/Maximum_position.png" decoding="async"
-                    style="max-width: 100%">
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-            </div>
-        </div>
-    </div>
-</div>
