--- static/help/vert_pos.html
@@ -1,25 +1,0 @@
-<!-- Modal -->
-<div class="modal fade" id="vert_pos_modal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-    <div class="modal-dialog modal-lg">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="exampleModalLabel">Vertical position</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <div class="modal-body">
-                <p>This position is obtained when the pendulum (ball + string) is in the
-                    perfectly vertical position. For that, slowly move the shovel in the backwards direction (towards
-                    the origin or micro-switch) until it has passed the vertical pendulum position. Wait for the ball to
-                    stop totally. Once the ball is fully stopped, approach the shovel in the direction of the ball until
-                    it touches the ball but without displacing it. Record the absolute shovel position by reading the
-                    value on the metric scale fixed on the launcher with the help of the indicator attached to the
-                    shovel.</p>
-                <br><img alt="Maximum position" src="/static/images/help/Vertical_position.png" decoding="async"
-                    style="max-width: 100%">
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-            </div>
-        </div>
-    </div>
-</div>
